backend/requirements.txt
fastapi
uvicorn
rembg
Pillow
python-multipart
openai
python-dotenv
aiohttp
onnxruntime
sqlmodel
bcrypt
pydantic[email]
python-jose[cryptography]
pydantic-settings
torch
peft
transformers
diffusers
simple-aesthetics-predictor
